Add accelerometer support.
The app now acts like a volume pedal: when the iPod is flat on the ground, the background is black, which would signify zero attenuation on the mic signal. If you tilt the iPod back toward you, the background turns redder, signifying some attenuation on the mic signal.

It would be nice to get smooth animations between the colors going, but I can't seem to make that work right now.