A little bit of cleanup...
I had hoped I could get the remote i/o audio unit to play stereo, but it seems to only want to handle mono.  I'm going to keep working on this, since I don't (yet) see any other way to do things.  Basically, the problem seems to be that since the recording/microphone is mono, the playback also has to be mono.  

I thought maybe I could work around this using an AUGraph and two separate audio units (one for recording and one for playback), but the only kind of unit available is remote i/o which is an "output unit", and only one output unit can exist in a graph at a time.  I then thought maybe I could use a subgraph to work around this limitation, but it appears that subgraphs aren't available on iphone.  Grr...

If none of my experiments work, one thing to try would be mixing a mono remote i/o unit (for recording) with a stereo AudioQueue (for playback), but I have no idea if/how that's possible.