-fixed things so that AudioButtonTestViewController.xib is now loaded properly and things don't crash anymore.  This held the key: http://discussions.apple.com/thread.jspa?threadID=1598422
Now there is an important question of when exactly the audio engine should be initialized and where in the project it should live.  It needs to be at some higher level than AudioButtonTestViewController, since the way things are now, audio can't happen until the first time that view is loaded.

-added project settings to allow app to be run on device (signing stuff).