I added the "circle touches" functionality to iDiMP's MainView.  For some reason, it kept crashing on me when trying to access the touchPoints NSMutableArray until I added a retain call after the call to create the array.  I don't understand why this was necessary, since it wasn't there in the original CircleTouches. :(  Now that I've added the retain call, maybe a release is necessary somewhere too??
Another weird thing that has been happening for me since before I added the circle touches functionality is that the "i" icon you touch to flip to the other side of the view is very temperamental.  It often doesn't respond at all.  I don't think it's the way I'm touching it, since I don't have this problem in any other app.  A bit of a mystery at the moment...